style: remove [allow(unused)] from all visitor
