+++ frontend/irgen/src/lib.rs
@@ -1,3 +1,4 @@
+mod impls;
 mod symbol_table;
 mod utils;
 mod visitor;
