+++ frontend/irgen/src/visitor.rs
@@ -1,186 +1,31 @@
-#![allow(unused)]
-
-use std::{
-	cmp::max,
-	collections::{HashMap, HashSet},
-};
+use std::collections::HashMap;
 
 use ast::{tree::*, Visitor};
 use attr::Attrs;
-use llvm::{
-	llvmop::{ArithOp, CompKind, CompOp, ConvertOp},
-	llvmvar::{self},
-	Value,
-	VarType::*,
-	*,
-};
+use llvm::{llvmop::*, Value, VarType::*, *};
 use rrvm::{
-	cfg::{link_basic_block, link_cfg, CFG},
-	program::{LlvmFunc, LlvmProgram, RrvmProgram},
+	cfg::link_cfg,
+	program::{LlvmFunc, LlvmProgram},
 	LlvmCFG,
 };
-use rrvm_symbol::{manager::SymbolManager, FuncSymbol, Symbol, VarSymbol};
-use utils::{
-	errors::Result,
-	Label,
-	SysycError::{LlvmvGenError, TypeError},
-};
-use value::{
-	calc_type::{to_rval, type_binaryop},
-	BType, BinaryOp, FuncRetType, FuncType, UnaryOp, VarType,
-};
+use rrvm_symbol::{FuncSymbol, VarSymbol};
+use utils::{errors::Result, Label, SysycError::TypeError};
+use value::{BinaryOp, FuncRetType, UnaryOp, VarType};
 
-use crate::{
-	symbol_table::{SymbolTable, Table},
-	utils::*,
-};
+use crate::{symbol_table::SymbolTable, utils::*};
 
 pub struct IRGenerator {
-	program: LlvmProgram,
-	stack: Vec<(LlvmCFG, Option<Value>, Option<Value>)>,
-	mgr: TempManager,
-	total: i32,
-	symbol_table: SymbolTable,
-	ret_type: FuncRetType,
+	pub program: LlvmProgram,
+	pub stack: Vec<(LlvmCFG, Option<Value>, Option<Value>)>,
+	pub mgr: TempManager,
+	pub total: i32,
+	pub symbol_table: SymbolTable,
+	pub ret_type: FuncRetType,
 }
 
 impl Default for IRGenerator {
 	fn default() -> Self {
 		Self::new()
-	}
-}
-
-impl IRGenerator {
-	pub fn new() -> Self {
-		Self {
-			program: RrvmProgram::new(),
-			stack: Vec::new(),
-			total: 0,
-			mgr: TempManager::new(),
-			symbol_table: SymbolTable::new(),
-			ret_type: FuncRetType::Void,
-		}
-	}
-	pub fn to_rrvm(mut self, program: &mut Program) -> Result<LlvmProgram> {
-		program.accept(&mut self)?;
-		Ok(self.program)
-	}
-	fn type_conv(
-		&mut self,
-		value: Value,
-		target: llvm::VarType,
-		cfg: &mut LlvmCFG,
-	) -> Value {
-		use llvmop::ConvertOp::*;
-		if target == value.get_type() {
-			return value;
-		}
-		let (from_type, to_type, op) = match target {
-			I32 => (F32, I32, Float2Int),
-			F32 => (I32, F32, Int2Float),
-			_ => unreachable!(),
-		};
-		match (target, &value) {
-			(F32, Value::Int(v)) => Value::Float(*v as f32),
-			(I32, Value::Float(v)) => Value::Int(*v as i32),
-			(_, Value::Temp(temp)) => {
-				let target = self.mgr.new_temp(to_type, false);
-				let instr = Box::new(ConvertInstr {
-					op,
-					target: target.clone(),
-					from_type,
-					lhs: temp.clone().into(),
-					to_type,
-				});
-				cfg.get_exit().borrow_mut().push(instr);
-				target.into()
-			}
-			_ => unreachable!(),
-		}
-	}
-	fn solve(
-		&mut self,
-		val: Option<Value>,
-		addr: Option<Value>,
-		cfg: &mut LlvmCFG,
-	) -> Value {
-		match val {
-			Some(value) => value,
-			None => {
-				let var_type = addr.as_ref().unwrap().deref_type();
-				let temp = self.mgr.new_temp(var_type, false);
-				let instr = Box::new(LoadInstr {
-					target: temp.clone(),
-					var_type,
-					addr: addr.unwrap(),
-				});
-				cfg.get_exit().borrow_mut().push(instr);
-				temp.into()
-			}
-		}
-	}
-	fn new_cfg(&mut self) -> LlvmCFG {
-		let out = CFG::new(self.total);
-		self.total += 1;
-		out
-	}
-	pub fn fold_cfgs(&mut self, cfgs: Vec<LlvmCFG>) -> LlvmCFG {
-		cfgs
-			.into_iter()
-			.reduce(|mut acc, mut v| {
-				link_cfg(&mut acc, &mut v);
-				acc.append(v);
-				acc
-			})
-			.unwrap_or_else(|| self.new_cfg())
-	}
-	pub fn if_then_else(
-		&mut self,
-		mut cond: LlvmCFG,
-		cond_val: Value,
-		mut cfg1: LlvmCFG,
-		diff1: Table,
-		mut cfg2: LlvmCFG,
-		diff2: Table,
-	) -> LlvmCFG {
-		let mut exit = self.new_cfg();
-		let keys = diff1
-			.keys()
-			.chain(diff2.keys())
-			.cloned()
-			.collect::<HashSet<_>>()
-			.into_iter();
-		fn get_val(id: i32, now: &Table, default: &SymbolTable) -> Value {
-			now.get(&id).map_or_else(|| default.get(&id), |v| v.clone())
-		}
-		for key in keys {
-			let val1 = get_val(key, &diff1, &self.symbol_table);
-			let val2 = get_val(key, &diff1, &self.symbol_table);
-			let var_type = val1.get_type();
-			let temp = self.mgr.new_temp(var_type, false);
-			let instr = PhiInstr {
-				target: temp.clone(),
-				var_type,
-				source: vec![(val1, cfg1.exit_label()), (val2, cfg2.exit_label())],
-			};
-			exit.get_exit().borrow_mut().push_phi(instr);
-			self.symbol_table.set(key, temp.into());
-		}
-		let instr = Box::new(JumpCondInstr {
-			var_type: I32,
-			cond: cond_val,
-			target_true: cfg1.entry_label(),
-			target_false: cfg2.entry_label(),
-		});
-		cond.get_exit().borrow_mut().set_jump(Some(instr));
-		link_cfg(&mut cond, &mut cfg1);
-		link_cfg(&mut cond, &mut cfg2);
-		link_cfg(&mut cfg1, &mut exit);
-		link_cfg(&mut cfg2, &mut exit);
-		cond.append(cfg1);
-		cond.append(cfg2);
-		cond.append(exit);
-		cond
 	}
 }
 
@@ -261,16 +106,12 @@ impl Visitor for IRGenerator {
 		self.stack.push((cfg, None, None));
 		Ok(())
 	}
-	fn visit_init_val_list(&mut self, node: &mut InitValList) -> Result<()> {
+	fn visit_init_val_list(&mut self, _node: &mut InitValList) -> Result<()> {
 		// TODO: solve init_val_list
 		todo!("I don't know how to solve this");
-		for val in node.val_list.iter_mut() {
-			val.accept(self)?;
-		}
-		Ok(())
 	}
 	fn visit_variable(&mut self, node: &mut Variable) -> Result<()> {
-		let mut now: LlvmCFG = self.new_cfg();
+		let now: LlvmCFG = self.new_cfg();
 		let symbol: VarSymbol = node.get_attr("symbol").unwrap().into();
 		let temp = self.symbol_table.get(&symbol.id);
 		if !symbol.var_type.is_array() {
@@ -281,24 +122,22 @@ impl Visitor for IRGenerator {
 		Ok(())
 	}
 	fn visit_literal_int(&mut self, node: &mut LiteralInt) -> Result<()> {
-		let mut now: LlvmCFG = self.new_cfg();
+		let now: LlvmCFG = self.new_cfg();
 		self.stack.push((now, Some(node.value.into()), None));
 		Ok(())
 	}
 	fn visit_literal_float(&mut self, node: &mut LiteralFloat) -> Result<()> {
-		let mut now: LlvmCFG = self.new_cfg();
+		let now: LlvmCFG = self.new_cfg();
 		self.stack.push((now, Some(node.value.into()), None));
 		Ok(())
 	}
 	fn visit_binary_expr(&mut self, node: &mut BinaryExpr) -> Result<()> {
 		use BinaryOp::*;
-		node.lhs.accept(self);
+		node.lhs.accept(self)?;
 		let (mut lcfg, lhs_val, lhs_addr) = self.stack.pop().unwrap();
 		self.symbol_table.push();
-		node.rhs.accept(self);
+		node.rhs.accept(self)?;
 		let (mut rcfg, rhs_val, rhs_addr) = self.stack.pop().unwrap();
-		let lhs_type: VarType = node.lhs.get_attr("type").unwrap().into();
-		let rhs_type: VarType = node.lhs.get_attr("type").unwrap().into();
 		let type_t: VarType = node.get_attr("type").unwrap().into();
 		let var_type = type_convert(&type_t);
 		let (cfg, ret_val, ret_addr) = match node.op {
@@ -523,22 +362,20 @@ impl Visitor for IRGenerator {
 		self.stack.push((cfg, None, None));
 		Ok(())
 	}
-	fn visit_while(&mut self, node: &mut While) -> Result<()> {
+	fn visit_while(&mut self, _node: &mut While) -> Result<()> {
+		// TODO: while
 		todo!();
-		node.cond.accept(self)?;
-		node.body.accept(self)?;
-		Ok(())
 	}
-	fn visit_continue(&mut self, node: &mut Continue) -> Result<()> {
+	fn visit_continue(&mut self, _node: &mut Continue) -> Result<()> {
 		/*
 		 这玩意本质是 goto 啊，咋处理来着
 		*/
+		// TODO: continue
 		todo!();
-		Ok(())
 	}
-	fn visit_break(&mut self, node: &mut Break) -> Result<()> {
+	fn visit_break(&mut self, _node: &mut Break) -> Result<()> {
+		// TODO: break
 		todo!();
-		Ok(())
 	}
 	fn visit_return(&mut self, node: &mut Return) -> Result<()> {
 		if let Some(val) = &mut node.value {
