+++ optimizer/Cargo.toml
@@ -6,4 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+utils = { path = "../utils" }
 rrvm = { path = "../utils/rrvm" }
+llvm = { path = "../utils/llvm" }
