+++ optimizer/src/dead_code/mod.rs
@@ -1,0 +1,3 @@
+pub mod impls;
+
+pub struct RemoveDeadCode {}
