+++ optimizer/src/impls.rs
@@ -1,12 +1,13 @@
 use rrvm::program::LlvmProgram;
+use utils::errors::Result;
 
-use crate::{BasicOptimizer, RrvmOptimizer};
+use crate::{dead_code::RemoveDeadCode, BasicOptimizer, RrvmOptimizer};
 
 impl RrvmOptimizer for BasicOptimizer {
 	fn new() -> Self {
 		Self {}
 	}
-	fn apply(self, _program: LlvmProgram) -> LlvmProgram {
-		todo!()
+	fn apply(self, program: &mut LlvmProgram) -> Result<()> {
+		RemoveDeadCode::new().apply(program)
 	}
 }
