+++ optimizer/src/lib.rs
@@ -1,9 +1,12 @@
-use rrvm::program::LlvmProgram;
+mod dead_code;
 pub mod impls;
+
+use rrvm::program::LlvmProgram;
+use utils::errors::Result;
 
 pub trait RrvmOptimizer {
 	fn new() -> Self;
-	fn apply(self, program: LlvmProgram) -> LlvmProgram;
+	fn apply(self, program: &mut LlvmProgram) -> Result<()>;
 }
 
 pub struct BasicOptimizer {}
