+++ utils/llvm/src/parser.rs
@@ -1,5 +1,3 @@
-#![allow(unused)]
-
 use pest::Parser;
 use pest_derive::Parser;
 use utils::{errors::Result, SysycError::LlvmSyntexError};
@@ -9,7 +7,7 @@ use utils::{errors::Result, SysycError::LlvmSyntexError};
 struct IrParser;
 
 pub fn parse(str: &str) -> Result<()> {
-	let progam = IrParser::parse(Rule::Program, str)
+	let _progam = IrParser::parse(Rule::Program, str)
 		.map_err(|e| LlvmSyntexError(e.to_string()))?;
 	todo!()
 }
